Haven't Pushed Changes. Oops.
This change creates a Tabletop save file when the user hits save. The
save file is currently using a default image from imgur to generate a
deck of cards. Next step is to upload the image the user creates to
imgur, and use that as a the image for the save file instead.